+++ src/services/cart/index.js
@@ -1,7 +1,7 @@
 import { addProductToCart, addListProductToCart } from './add'
 import deleteOneProductFromCart from './delete'
 import updateFieldQuantiyCartItem from './update'
-import listingProductInCart from './show'
+import listingProductInCart from './get'
 
 export {
   addProductToCart,
