+++ src/services/wishlist/index.js
@@ -1,6 +1,6 @@
 import { addProductToWishlist, addListProductToWishlist } from './add'
 import deleteOneProductFromWishlist from './delete'
-import listingProductInWishlist from './show'
+import listingProductInWishlist from './get'
 
 export {
   addProductToWishlist,
